requirements.txt
azure-core==1.30.1
azure-identity==1.16.0
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
matplotlib==3.8.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
registry==0.4.2
requests==2.31.0
setuptools==68.0.0
simpy==4.1.1
sympy==1.12
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
